+++ app/src/main/java/dev/yasan/metro/tehran/ui/theme/Color.kt
@@ -2,12 +2,7 @@ package dev.yasan.metro.tehran.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val themePrimary = Color(0xFFb2dfdb)
-val themePrimaryDark = Color(0xFF82ada9)
-val themePrimaryLight = Color(0xFFe5ffff)
-val onThemePrimary = Color.Black
-
-val themeAccent = Color(0xFFffd600)
-val themeAccentDark = Color(0xFFc7a500)
-val themeAccentLight = Color(0xFFffff52)
-val onThemeAccent = Color.Black
+val themePrimary = Color(0xFF333333)
+val themePrimaryDark = Color(0xFF0c0c0c)
+val themePrimaryLight = Color(0xFF5c5c5c)
+val onThemePrimary = Color.White