+++ app/src/main/java/dev/yasan/metro/tehran/ui/theme/Dimensions.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Stable
 import androidx.compose.ui.unit.dp
 
 private val dimenGrid = 8.dp
-val dimenDivider = 1.dp
+val dimenDivider = 2.dp
 
 /**
  * @return The base grid size multiplied by [multiplier].